move the generated enum into autogen/types.ts
