+++ src/actions/parse.ts
@@ -37,38 +37,6 @@ const configSchema = z.object({
   actions: z.record(z.string(), providerSchema),
 });
 
-// This will be populated dynamically during type generation
-export enum ProviderName {
-    GENERIC = "generic",
-    ASANA = "asana",
-    SLACK = "slack",
-    MATH = "math",
-    CONFLUENCE = "confluence",
-    JIRA = "jira",
-    KANDJI = "kandji",
-    GOOGLEMAPS = "googlemaps",
-    BING = "bing",
-    ZENDESK = "zendesk",
-    LINKEDIN = "linkedin",
-    X = "x",
-    MONGO = "mongo",
-    SNOWFLAKE = "snowflake",
-    OPENSTREETMAP = "openstreetmap",
-    NWS = "nws",
-    FIRECRAWL = "firecrawl",
-    RESEND = "resend",
-    GOOGLEOAUTH = "googleOauth",
-    GOOGLEMAIL = "googlemail",
-    GONG = "gong",
-    FINNHUB = "finnhub",
-    LOOKER = "looker",
-    ASHBY = "ashby",
-    SALESFORCE = "salesforce",
-    MICROSOFT = "microsoft",
-    GITHUB = "github",
-    NOTION = "notion"
-}
-
 const authParamsSchemaStr = `
 z.object({
     authToken: z.string().optional(),
@@ -192,18 +160,16 @@ async function generateTypes({
   const templatesFile = project.createSourceFile(outputPath, "", { overwrite: true });
   const typesFile = project.createSourceFile(templatesOutputPath, "", { overwrite: true });
 
-  // Update the ProviderName enum in parse.ts
-  const parseFile = project.addSourceFileAtPath("src/actions/parse.ts");
-  const providerEnum = parseFile.getEnum("ProviderName");
-  if (providerEnum) {
-    providerEnum.getMembers().forEach(member => member.remove());
-    Object.keys(parsedConfig.actions).forEach(providerName => {
-      providerEnum.addMember({
+  // Set the ProviderName enum based on the schema providers
+  typesFile.addEnum({
+    name: "ProviderName",
+    members: Object.keys(parsedConfig.actions).map(providerName => 
+      ({
         name: providerName.toUpperCase().replace(/-/g, '_'),
         value: providerName,
-      });
-    });
-  }
+      })
+    ),
+  }).setIsExported(true);
 
   // Add imports
   templatesFile.addImportDeclaration({
@@ -288,7 +254,6 @@ async function generateTypes({
   // Save the generated TypeScript file
   await templatesFile.save();
   await typesFile.save();
-  await parseFile.save();
 }
 
 generateTypes({});
