+++ src/index.ts
@@ -3,7 +3,7 @@
 export { runAction, getActions, type ActionGroupsReturn } from "./app";
 export { ACTION_GROUPS, type ActionGroups } from "./actions/groups";
 export { ActionMapper } from "./actions/actionMapper";
-export { ActionTemplate, ProviderName } from "./actions/parse";
+export { ActionTemplate } from "./actions/parse";
 
 export * from "./actions/autogen/templates";
 export * from "./actions/autogen/types";
